requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.2
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
Markdown==3.8
openai==1.75.0
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2

# Production dependencies
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
django-widget-tweaks==1.5.0

# AWS S3 storage
boto3>=1.34.0
django-storages==1.14.6
